app/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.23
PyMySQL==1.1.0
bcrypt==4.0.1
redis==5.0.1
python-magic==0.4.27
Pillow==10.2.0
gunicorn
gevent-websocket
cryptography
gevent==23.9.1
pytest==7.4.3
fakeredis==2.20.0
structlog==23.2.0
